Remove old gnosis tx service test
